Move Mini Cart block component into its own file
